✨ Feat: 유효하지 않은 문자열 판단 메서드 추가
